# openapi.yml
openapi: 3.0.3
info:
  title: SIGC - Sistema de Gestión de Citas Médicas
  description: API completa para gestión de especialidades, doctores, citas y usuarios
  version: 1.0.0
  contact:
    name: Soporte SIGC
    email: soporte@sigc.com
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo
  - url: https://api.sigc.com
    description: Servidor de producción
...
paths:
...
  /doctores:
    get:
      tags:
        - Doctores
      summary: Listar todos los doctores
      description: Retorna una lista de doctores disponibles con filtros opcionales
      parameters:
        - name: especialidad
          in: query
          schema:
            type: string
          description: Filtrar por especialidad
          example: Cardiología
        - name: disponibles
          in: query
          schema:
            type: boolean
          description: Mostrar solo doctores disponibles
      responses:
        '200':
          description: Lista de doctores obtenida
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DoctorResponse'
              example:
                - idDoctor: 1
                  nombre: Dr. Carlos López
                  especialidad: Cardiología
                  telefono: "555001111"
                  email: carlos@sigc.com
                  disponible: true

    post:
      tags:
        - Doctores
      summary: Registrar nuevo doctor
      description: Registra un nuevo doctor en el sistema (requiere permisos de admin)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DoctorRequest'
      responses:
        '201':
          description: Doctor registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DoctorResponse'
        '401':
          description: No autorizado
        '400':
          description: Datos inválidos
  /doctores/{id}:
    get:
      tags:
        - Doctores
      summary: Obtener doctor por ID
      description: Retorna los datos de un doctor específico
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Doctor encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DoctorResponse'
        '404':
          description: Doctor no encontrado
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido al hacer login
  schemas:
...
    DoctorResponse:
      type: object
      properties:
        idDoctor:
          type: integer
          example: 1
        nombre:
          type: string
          example: Dr. Carlos López
        especialidad:
          type: string
          example: Cardiología
        telefono:
          type: string
          example: "555001111"
        email:
          type: string
          example: carlos@sigc.com
        disponible:
          type: boolean
          example: true
        horarioAtencion:
          type: string
          example: "09:00 - 18:00"
    DoctorRequest:
      type: object
      required:
        - nombre
        - especialidad
        - telefono
        - email
      properties:
        nombre:
          type: string
          example: Dr. Carlos López
        especialidad:
          type: string
          example: Cardiología
        telefono:
          type: string
          pattern: '^\d{9}$'
          example: "555001111"
        email:
          type: string
          format: email
          example: carlos@sigc.com
        horarioAtencion:
          type: string
          example: "09:00 - 18:00"
...